extract types
