+++ chain-extensions/types/signing/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "signing-chain-extension-types"
+version = "1.1.0"
+license = "Apache-2.0"
+description = "Types definitions for signing chain-extension"
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+frame-support = { workspace = true }
+parity-scale-codec = { workspace = true }
+scale-info = { workspace = true }
+sp-core = { workspace = true }
+sp-runtime = { workspace = true }
+
+[features]
+default = ["std"]
+std = [
+	"parity-scale-codec/std",
+	"frame-support/std",
+	"scale-info/std",
+	"sp-core/std",
+	"sp-runtime/std",
+]
